solving the bundle problem
Currently, a gem that uses structured_changelog can't be bundled until
structured_changelog exists on contributors' machines, which is bad, as
it gives any gem that uses it an additional install step. To get around
that, we're recommending people uses `rake changelog:sync` when they
need to update their gem's VERSION constant. This sets that constant to
the latest Changelog release version. We've also added `validate` and
`sync` as prereq tasks to `rake release`, so your VERSION constant gets
automatically updated when you go to release. This keeps the number of
release steps the same, and making gems using structured_changelog as
easy to install as anything else.