had to make a changelog:release task :(
I would have liked to just wrap existing tasks for this, but no. Here's
why:

Merely requiring `bundler/gem_tasks` instantiates & caches a
gemspec. This means that our changes to the filesystem won't change the
version bundler thinks our gem is. We could instantiate another gemspec
and attach it, or we could try modifying the gemspec in situ, but both
would be extremely brittle. What we do instead, is start another
process, and because that process doesn't have the state ours does
already, our changes to the filesystem will be picked up, and we'll
actually cut a release of the new version.